fixed. #18
Changed how to manage the resulting file

+++ lib/create_chart.sh
@@ -3,7 +3,7 @@
 #
 #         FILE: create_chart.sh
 #
-#        USAGE: create_chart.sh [-o directory] [-c cpuList] [-i i/fList] [-d deviceList] [-m mountedList] [-h]
+#        USAGE: create_chart.sh [-o directory] [-d delimiter] [-h] measure-map
 #
 #  DESCRIPTION: Create Visualize Tool
 #
@@ -19,16 +19,21 @@ function usage() {
 cat << EOF
 Usage:
 
-  ${0} [-o directory] [-c cpuList] [-i i/fList] [-d deviceList] [-m mountedList] [-h]
+  ${0} [-o directory] [-d delimiter] [-h] measure-map
 
     -o <arg> : Specify results directory
                Default : '\$(cd \$(dirname \$0);pwd)/result-\`date +%Y%m%d%H%M%S\`'
                  For example, '$(cd $(dirname $0);pwd)/result-`date +%Y%m%d%H%M%S`'
-    -c <arg> : CPU List
-    -i <arg> : Interfaces List
-    -d <arg> : Device List
-    -m <arg> : Mounted List
+    -d <arg> : measure-map delimiter
+               Default : ':'
     -h       : Get help
+    measure-map
+             : Format
+               <measure-type>:<file-path>:<name>
+
+                 measure-type : Measurement type
+                 file-path    : Relative path from the results
+                 name         : Measurement item name
 
 EOF
 exit 0
@@ -60,78 +65,86 @@ CONTAINER_IDX=0
 #-------------------------------------------------------------------------------
 # Parameter check
 #-------------------------------------------------------------------------------
-while getopts "o:c:i:d:m:h" OPT; do
+while getopts "o:d:h" OPT; do
   case ${OPT} in
     o) RESULT_DIR="${OPTARG}";;
-    c) LIST_CPU="${OPTARG}";;
-    i) LIST_IF="${OPTARG}";;
-    d) LIST_DEV="${OPTARG}";;
-    m) LIST_MNT="${OPTARG}";;
+    d) MAP_DELIMITER="${OPTARG}";;
     h|:|\?) usage;;
   esac
 done
-
 shift $(( $OPTIND - 1 ))
 
-test ! -d ${RESULT_DIR} && mkdir -p ${RESULT_DIR}
+MEASURE_MAP=${1}
+MAP_DELIMITER=${MAP_DELIMITER:-:}
 
-#cp -p ${VIEW_DIR}/html/measure.html ${RESULT_DIR}
-#cp -p ${VIEW_DIR}/html/measure.js   ${RESULT_DIR}
-#cp -p ${VIEW_DIR}/html/measure.css  ${RESULT_DIR}
+if [ ! -f ${MEASURE_MAP} ]; then
+  echo "measure-map does not exist." >&2
+  exit 1
+fi
+
+test ! -d ${RESULT_DIR} && mkdir -p ${RESULT_DIR}
 cp -pr ${VIEW_DIR}/* ${RESULT_DIR}
 
 #-------------------------------------------------------------------------------
 # Add a button for viewing memory measure results
 #-------------------------------------------------------------------------------
-add_button 'MEMORY' '<input class="btn_mem" type="button" target="UUID" src="data/memory.csv" value="Memory" />'
+grep "^memory${MAP_DELIMITER}" ${MEASURE_MAP} | while read line
+do
+  path="`echo ${line} | cut -d ${MAP_DELIMITER} -f2`"
+  name="`echo ${line} | cut -d ${MAP_DELIMITER} -f3-`"
+  add_button 'MEMORY' `echo '<input class="btn_mem" type="button" target="UUID" src="SRC" value="VAL" />' | sed "s#SRC#${path}#g" | sed "s#VAL#${name}#g"`
+done
 
 #-------------------------------------------------------------------------------
 # Add a button for viewing CPU measure results
 #-------------------------------------------------------------------------------
-add_button 'CPU' '<input class="btn_cpu" type="button" target="UUID" src="data/cpu.all.csv" value="all" />'
-for i in ${LIST_CPU[@]}
+grep "^cpu${MAP_DELIMITER}" ${MEASURE_MAP} | while read line
 do
-  add_button 'CPU' `echo '<input class="btn_cpu" type="button" target="UUID" src="data/cpu.###.csv" value="###" />' | sed "s/###/$i/g"`
+  path="`echo ${line} | cut -d ${MAP_DELIMITER} -f2`"
+  name="`echo ${line} | cut -d ${MAP_DELIMITER} -f3-`"
+  add_button 'CPU' `echo '<input class="btn_cpu" type="button" target="UUID" src="SRC" value="VAL" />' | sed "s#SRC#${path}#g" | sed "s#VAL#${name}#g"`
 done
 
 #-------------------------------------------------------------------------------
 # Add a button for viewing Load Average measure results
 #-------------------------------------------------------------------------------
-add_button 'LOAD_AVERAGE' '<input class="btn_loadavg" type="button" target="UUID" src="data/loadavg.csv" value="LoadAverage" />'
+grep "^loadavg${MAP_DELIMITER}" ${MEASURE_MAP} | while read line
+do
+  path="`echo ${line} | cut -d ${MAP_DELIMITER} -f2`"
+  name="`echo ${line} | cut -d ${MAP_DELIMITER} -f3-`"
+  add_button 'LOAD_AVERAGE' `echo '<input class="btn_loadavg" type="button" target="UUID" src="SRC" value="VAL" />' | sed "s#SRC#${path}#g" | sed "s#VAL#${name}#g"`
+done
 
 #-------------------------------------------------------------------------------
 # Add a button for viewing Network Traffic measure results
 #-------------------------------------------------------------------------------
-for i in ${LIST_IF[@]}
+grep "^network${MAP_DELIMITER}" ${MEASURE_MAP} | while read line
 do
-  add_button 'NETWORK' `echo '<input class="btn_network" type="button" target="UUID" src="data/network.###.csv" value="###" />' | sed "s/###/$i/g"`
+  path="`echo ${line} | cut -d ${MAP_DELIMITER} -f2`"
+  name="`echo ${line} | cut -d ${MAP_DELIMITER} -f3-`"
+  add_button 'NETWORK' `echo '<input class="btn_network" type="button" target="UUID" src="SRC" value="VAL" />' | sed "s#SRC#${path}#g" | sed "s#VAL#${name}#g"`
 done
 
 #-------------------------------------------------------------------------------
 # Add a button for viewing Disk IO measure results
-#-------------------------------------------------------------------------------
-for i in ${LIST_DEV[@]}
-do
-  add_button 'DISK_IO' `echo '<input class="btn_diskio" type="button" target="UUID" src="data/diskio.###.csv" value="###" />' | sed "s/###/$i/g"`
-done
-
-#-------------------------------------------------------------------------------
 # Add a button for viewing Disk Util measure results
 #-------------------------------------------------------------------------------
-for i in ${LIST_DEV[@]}
+grep "^device${MAP_DELIMITER}" ${MEASURE_MAP} | while read line
 do
-  add_button 'DISK_UTIL' `echo '<input class="btn_diskutil" type="button" target="UUID" src="data/diskio.###.csv" value="###" />' | sed "s/###/$i/g"`
+  path="`echo ${line} | cut -d ${MAP_DELIMITER} -f2`"
+  name="`echo ${line} | cut -d ${MAP_DELIMITER} -f3-`"
+  add_button 'DISK_IO' `echo '<input class="btn_diskio" type="button" target="UUID" src="SRC" value="VAL" />' | sed "s#SRC#${path}#g" | sed "s#VAL#${name}#g"`
+  add_button 'DISK_UTIL' `echo '<input class="btn_diskutil" type="button" target="UUID" src="SRC" value="VAL" />' | sed "s#SRC#${path}#g" | sed "s#VAL#${name}#g"`
 done
 
 #-------------------------------------------------------------------------------
 # Add a button for viewing Disk Usage measure results
 #-------------------------------------------------------------------------------
-for i in ${LIST_MNT[@]}
+grep "^mount${MAP_DELIMITER}" ${MEASURE_MAP} | while read line
 do
-  filename=`echo 'diskuse.%%%.csv' | sed "s/%%%/${i//\//\\_S_}/g"`
-  add_button 'DISK_USAGE' `echo '<input class="btn_diskuse" type="button" target="UUID" src="data/FILE" value="%%%" />' | \
-    sed "s/FILE/${filename}/g" | \
-    sed "s/%%%/${i//\//\\/}/g"`
+  path="`echo ${line} | cut -d ${MAP_DELIMITER} -f2`"
+  name="`echo ${line} | cut -d ${MAP_DELIMITER} -f3-`"
+  add_button 'DISK_USAGE' `echo '<input class="btn_diskuse" type="button" target="UUID" src="SRC" value="VAL" />' | sed "s#SRC#${path}#g" | sed "s#VAL#${name}#g"`
 done
 
 exit 0
